refactor : request response 분리
