+++ src/main/java/bc1/gream/domain/admin/dto/request/AdminProductRequestDto.java
@@ -1,8 +1,8 @@
-package bc1.gream.domain.admin.dto;
+package bc1.gream.domain.admin.dto.request;
 
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.Builder;
 
-@JsonIgnoreProperties
+@Builder
 public record AdminProductRequestDto(
     String brand,
     String name,
