+++ src/main/java/bc1/gream/domain/admin/dto/response/AdminProductResponseDto.java
@@ -1,4 +1,4 @@
-package bc1.gream.domain.admin.dto;
+package bc1.gream.domain.admin.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
